feat(infrastructure): improved reliability and performance of build
