+++ firebase.json
@@ -100,6 +100,16 @@
       "source": "/docs/*",
       "destination": "/api/",
       "type": "301"
+    },
+    {
+      "source": "/api/",
+      "destination": "/api/v18/documentation/",
+      "type": "301"
+    },
+    {
+      "source": "/api/v18/",
+      "destination": "/api/v18/documentation/",
+      "type": "301"
     }
   ]
 }
