+++ package.json
@@ -36,22 +36,16 @@
     "gatsby-plugin-manifest": "^4.20.0",
     "gatsby-plugin-mdx": "^3.20.0",
     "gatsby-plugin-meta-redirect": "^1.1.1",
-    "gatsby-plugin-minify-html": "^1.0.4",
     "gatsby-plugin-offline-next": "^5.2.3",
     "gatsby-plugin-sass": "^5.20.0",
-    "gatsby-plugin-sharp": "^4.20.0",
     "gatsby-plugin-sitemap": "^5.20.0",
     "gatsby-plugin-svgr": "^3.0.0-beta.0",
     "gatsby-plugin-typescript": "^4.20.0",
     "gatsby-plugin-webfonts": "^2.2.2",
-    "gatsby-react-router-scroll": "^5.20.0",
     "gatsby-remark-autolink-headers": "^5.21.0",
-    "gatsby-remark-copy-linked-files": "^5.21.0",
-    "gatsby-remark-images": "^6.21.0",
     "gatsby-remark-prismjs": "^6.21.0",
     "gatsby-source-filesystem": "^4.20.0",
     "gatsby-theme-i18n": "^3.0.0",
-    "gatsby-transformer-remark": "^5.20.0",
     "gatsby-transformer-yaml": "^4.20.0",
     "highlight.js": "^11.6.0",
     "intersection-observer": "^0.12.2",
@@ -78,10 +72,9 @@
   "license": "MIT",
   "scripts": {
     "cleanup": "gatsby clean",
-    "pre-build": "node scripts/generateRedirects.js",
     "sync-api": "node scripts/syncApiDocs.js",
-    "build": "npm run pre-build && gatsby build",
-    "build-ci": "npm run cleanup && npm run pre-build && gatsby build --prefix-paths",
+    "build": "gatsby build",
+    "build-ci": "gatsby build --prefix-paths",
     "start": "gatsby develop",
     "format": "prettier --write \"**/*.{ts,tsx,js,scss}\" && npm run lint:fix",
     "format-check": "prettier --check \"**/*.{ts,tsx,js,scss}\"",
